Add automatic 802.3ad bonding support for servers.
Module must be initially enabled as most setups probably aren't using
bond interfaces and won't need the 30 second wait to scan for them.

Feature tries to tcpdump all non-wireless interfaces listening for LACP
packets concurrently to save time.

With a default timeout of 35 seconds, if the hook is enabled and notices
an LACP neighbour it'll note down their advertised MAC and make a
temporary bond interface pairing all interfaces which received LACPs
from the same advertised MAC (In attempt to avoid crisscrossing multiple
bonds). Worked well in testing with QEMU, to see how real hardware likes
it.

10_dhcp/hook has been modified to accommodate this change. Instead of
running blindly it now checks that interfaces aren't part of a bond or
bridge before adding them to a temporary /etc/dhclient.conf if it deems
them valid for dhcp. This prevents dhclient from trying to run on the
underlying bonded interfaces and causing issues.