Fix mount handler quirk with child rootfs datasets
Previously, datasets which existed under the desired root would
overmount the newroot mountpoint. In the case of Docker with the ZFS
storage engine enabled, this made it possible for the PC to boot into a
pulled docker image, accidentally softlocking on a login prompt where
the docker image has no valid credentials to progress further than that.

The main hook has been updated to better ignore datasets which don't
want to be mounted or have mountpoint set to legacy, opting to leave
those up to the host's /etc/fstab entry processing once its finished.

It will still mount child datasets of the rootfs where their mountpoint
is set to a full filesystem path.